Make amount of tokens in throttle! configurable
To implement "ping" for a throttle (to use it for rate limiting
without having to increment it), or to implement rate limiting
in arbitrary units.

It has two main use cases:

Ping: keep a throttle "alive" throughout multiple invocations,
some of which do cause it to increment and some don't.
Metered rate: increment by values above 1, for example by 1
for each item added to a cart or for every unit of payload sent
to some (expensive) remote service.