Use fractional time measurement in Redis lua
Redis returns time as a tuple. Since the computation of the leak
rate invariable involves fractions, it is very helpful to be able
to have more precision than 1 second. Having only 1-second precision
otherwise means that the behavior of the throttle would be dependent
on how much time has passed since the second rollover, which is not
correct: our leak rate is fractional and not "per second", so our
delta-T measurement should also use fractional time.