Add a "naked" LeakyBucket module (#28)
The throttle is very full featured but it also carries a lot
of stuff one doesn't necessarily need. There can be situations where
all you need is a leaky bucket and you want to measure its state,
in terms of tokens present. For example if you want to determine if
a specific item is "hot" without necessarily raising exceptions or
setting extra Redis keys.

Let's see if we can reduce the unit of functionality to just that,
and have a module with less configuration too.